# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Paperless-LLM API
  version: 1.0.0
  description: |
    REST API for the Paperless-LLM document processing system. 
    
    This API enables automated document processing workflows using Large Language Models (LLMs)
    to generate document metadata such as titles, tags, and summaries. It supports both automated
    and approval-based workflows with comprehensive job tracking, queue management, and step-level
    retry mechanisms.
    
    ## Key Features
    - **Job Submission**: Submit batch processing jobs for multiple documents
    - **Workflow Management**: Track job progress through multi-step workflows
    - **Approval System**: Review and approve/reject LLM-generated suggestions
    - **Queue Monitoring**: Real-time visibility into processing queues
    - **Prompt Customization**: Customize LLM prompts for different processing steps
    - **Health Monitoring**: Monitor system and service health status
    - **Retry Mechanism**: Automatic retry with exponential backoff for failed steps
    
    ## Architecture
    The system follows a clean architecture pattern with:
    - Domain-Driven Design principles
    - Step-based workflow orchestration
    - Transaction management for data consistency
    - RFC 7807 Problem Details for error responses
  contact:
    name: Paperless-LLM Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: API base path
...
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: List jobs with pagination
      description: Retrieve a paginated list of jobs, optionally filtered by job state
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          description: Maximum number of jobs to return (default 50, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor from a previous response
          schema:
            type: string
        - name: state
          in: query
          description: Filter jobs by state
          schema:
            $ref: '#/components/schemas/JobState'
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Jobs
      summary: Submit batch job
      description: |
        Submit one or more documents for processing. Each document can have multiple job types.
        Jobs are processed asynchronously through a multi-step workflow.
      operationId: submitJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchJobRequest'
            examples:
              single-document-single-job:
                summary: Single document, single job type
                value:
                  documents:
                    - documentId: "123"
                      jobTypes: ["automated"]
                      fields: ["title"]
              multiple-documents:
                summary: Multiple documents with different job types
                value:
                  documents:
                    - documentId: "123"
                      jobTypes: ["title"]
                      fields: ["title", "tags"]
                    - documentId: "456"
                      jobTypes: ["title", "tag"]
                      fields: ["title", "tags"]
              with-approval:
                summary: Job requiring approval
                value:
                  documents:
                    - documentId: "789"
                      jobTypes: ["approval"]
      responses:
        '201':
          description: Jobs submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /jobs/stats:
    get:
      tags:
        - Jobs
      summary: Get job statistics
      description: Returns count of jobs grouped by their current state
      operationId: getJobStats
      responses:
        '200':
          description: Job statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /jobs/{id}:
    get:
      tags:
        - Jobs
      summary: Get job details
      description: Retrieve detailed information about a specific job including its current status and document actions
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /jobs/{id}/steps:
    get:
      tags:
        - Jobs
      summary: Get job workflow steps
      description: Retrieve all workflow steps for a specific job including their execution status and retry information
      operationId: getJobSteps
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job workflow steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStepsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /jobs/{id}/audit-log:
    get:
      tags:
        - Jobs
      summary: Get job audit log
      description: Retrieve the complete audit trail for a specific job, including all step executions, retries, approvals, and state changes
      operationId: getJobAuditLog
      parameters:
        - name: id
          in: path
          required: true
          description: Job ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
              example:
                auditLog:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    jobId: "123e4567-e89b-12d3-a456-426614174000"
                    stepId: null
                    eventType: "JOB_CREATED"
                    eventTimestamp: "2024-01-15T10:00:00.000Z"
                    processingStartTime: null
                    processingEndTime: null
                    processingDurationMs: null
                    metadata:
                      documentId: "42"
                      jobType: "GENERATE_TITLE"
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    jobId: "123e4567-e89b-12d3-a456-426614174000"
                    stepId: "789e4567-e89b-12d3-a456-426614174000"
                    eventType: "STEP_COMPLETED"
                    eventTimestamp: "2024-01-15T10:00:15.000Z"
                    processingStartTime: "2024-01-15T10:00:10.000Z"
                    processingEndTime: "2024-01-15T10:00:15.000Z"
                    processingDurationMs: 5000
                    metadata:
                      stepType: "LLM_GENERATE_TITLE"
                      retryCount: 0
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /approvals:
    get:
      tags:
        - Approvals
      summary: List pending approvals
      description: Retrieve a paginated list of steps awaiting approval decisions
      operationId: listApprovals
      parameters:
        - name: limit
          in: query
          description: Maximum number of approvals to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor from a previous response
          schema:
            type: string
      responses:
        '200':
          description: List of pending approvals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /approvals/stats:
    get:
      tags:
        - Approvals
      summary: Get approval statistics
      description: Returns count of pending approval requests
      operationId: getApprovalStats
      responses:
        '200':
          description: Approval statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /approvals/{stepId}:
    post:
      tags:
        - Approvals
      summary: Make approval decision
      description: |
        Submit an approval decision (approve/reject) for a pending step.
        The decision must match one of the possibleDecisions from the approval item.
      operationId: makeApprovalDecision
      parameters:
        - name: stepId
          in: path
          required: true
          description: Step ID awaiting approval
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  description: Decision value (must match possibleDecisions)
                  example: approve
              required:
                - decision
            examples:
              approve:
                summary: Approve the proposed changes
                value:
                  decision: approve
              reject:
                summary: Reject the proposed changes
                value:
                  decision: reject
      responses:
        '200':
          description: Decision processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    WorkflowType:
      type: string
      enum:
        - approval
        - automated
      description: Type of workflow processing mode
    JobState:
      type: string
      enum:
        - pending
        - llm_processing
        - pending_approval
        - updating_document
        - completed
        - failed
        - rejected
      description: Current state of a job in its workflow
    StepType:
      type: string
      enum:
        - LLM_GENERATE_TITLE
        - LLM_GENERATE_FIELDS
        - LLM_GENERATE_TAGS
        - LLM_GENERATE_CORRESPONDENT
        - LLM_GENERATE_DOCUMENT_TYPE
        - LLM_GENERATE_CREATED_DATE
        - REQUIRE_APPROVAL
        - UPDATE_DOCUMENT
        - REMOVE_TAGS
      description: Type of workflow step
...
    BatchJobRequest:
      type: object
      properties:
        documents:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                description: ID of the document in Paperless-NGX
                example: "123"
              jobType:
                $ref: '#/components/schemas/WorkflowType'
              fields:
                  $ref: '#/components/schemas/LLMGenerateableFields'
            required:
              - documentId
              - jobType
              - fields
          minItems: 1
      required:
        - documents
    LLMGenerateableFields:
      type: array
      items:
        type: string
        enum: ["title", "tags", "correspondent", "document_type", "created_date"]


    # Response Schemas
    JobSubmissionResponse:
      type: object
      properties:
        submitted:
          type: integer
          description: Number of jobs successfully submitted
          example: 2
        jobs:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                example: "123"
              jobType:
                type: string
                example: title
              id:
                type: string
                format: uuid
                example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            required:
              - documentId
              - jobType
              - id
      required:
        - submitted
        - jobs
    JobResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        fields:
          $ref: '#/components/schemas/LLMGenerateableFields'
        documentId:
          type: string
          example: "123"
        jobType:
          $ref: '#/components/schemas/WorkflowType'
        status:
          $ref: '#/components/schemas/JobState'
        errorMessage:
          type: string
          nullable: true
          description: Error message if job failed
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-11T10:15:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-11T10:30:00Z"
        documentActions:
          type: array
          items:
            $ref: '#/components/schemas/DocumentAction'
      required:
        - id
        - documentId
        - jobType
        - status
        - createdAt
        - updatedAt
        - documentActions
    DocumentAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actionType:
          type: string
          description: Type of document action (e.g., 'update_title', 'add_tag')
          example: update_title
        oldValue:
          type: string
          nullable: true
          description: Previous value before the action
          example: "Untitled Document"
        newValue:
          type: string
          nullable: true
          description: New value after the action
          example: "Invoice 2026-05-11"
      required:
        - id
        - actionType
        - oldValue
        - newValue
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobResponse'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page of results
      required:
        - jobs
        - nextCursor
    JobStepsResponse:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/JobStep'
      required:
        - steps
    JobStep:
      type: object
      properties:
        stepId:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        stepType:
          $ref: '#/components/schemas/StepType'
        stepStatus:
          $ref: '#/components/schemas/StepStatus'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
        startedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-11T10:01:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-11T10:05:00Z"
        retryCount:
          type: integer
          description: Number of times this step has been retried
          minimum: 0
          example: 0
        retryAfter:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when retry will be attempted
          example: null
      required:
        - stepId
        - stepType
        - stepStatus
        - createdAt
        - retryCount
        - retryAfter
    AuditLogResponse:
      type: object
      properties:
        auditLog:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
      required:
        - auditLog
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        jobId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        stepId:
          type: string
          format: uuid
          nullable: true
          description: Step ID (null for job-level events)
          example: "789e4567-e89b-12d3-a456-426614174000"
        eventType:
          $ref: '#/components/schemas/AuditEventType'
        eventTimestamp:
          type: string
          format: date-time
          description: When the event occurred
          example: "2024-01-15T10:00:00.000Z"
        processingStartTime:
          type: string
          format: date-time
          nullable: true
          description: When processing started (for execution events)
          example: "2024-01-15T10:00:10.000Z"
        processingEndTime:
          type: string
          format: date-time
          nullable: true
          description: When processing ended (for execution events)
          example: "2024-01-15T10:00:15.000Z"
        processingDurationMs:
          type: integer
          nullable: true
          description: Processing duration in milliseconds
          minimum: 0
          example: 5000
        metadata:
          type: object
          nullable: true
          description: Event-specific metadata (varies by eventType)
          additionalProperties: true
      required:
        - id
        - jobId
        - eventType
        - eventTimestamp
    JobStats:
      type: object
      properties:
        pending:
          type: integer
          minimum: 0
          example: 5
        llmProcessing:
          type: integer
          minimum: 0
          example: 3
        pendingApproval:
          type: integer
          minimum: 0
          example: 2
        updatingDocument:
          type: integer
          minimum: 0
          example: 1
        completed:
          type: integer
          minimum: 0
          example: 42
        failed:
          type: integer
          minimum: 0
          example: 1
        rejected:
          type: integer
          minimum: 0
          example: 0
      required:
        - pending
        - llmProcessing
        - pendingApproval
        - updatingDocument
        - completed
        - failed
        - rejected
    ApprovalsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalItem'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page of results
      required:
        - items
        - nextCursor
    ApprovalItem:
      type: object
      properties:
        stepId:
          type: string
          format: uuid
          example: "c3d4e5f6-a7b8-9012-cdef-123456789012"
        jobId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        documentId:
          type: string
          example: "123"
        paperlessUrl:
          type: string
          format: uri
          description: URL to view the document in Paperless-NGX
          example: "http://paperless.local/documents/123"
        jobType:
          type: string
          example: title
        proposedActions:
          type: array
          items:
            type: object
            properties:
              actionType:
                type: string
                example: update_title
              oldValue:
                type: string
                example: "Untitled"
              newValue:
                type: string
                example: "Invoice 2026-05-11"
            required:
              - actionType
              - oldValue
              - newValue
        possibleDecisions:
          type: array
          items:
            type: string
          example: ["approve", "reject"]
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:15:00Z"
      required:
        - stepId
        - jobId
        - documentId
        - paperlessUrl
        - jobType
        - proposedActions
        - possibleDecisions
        - createdAt
    ApprovalStats:
      type: object
      properties:
        pendingCount:
          type: integer
          minimum: 0
          example: 3
      required:
        - pendingCount
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "about:blank"
        title:
          type: string
          description: Short, human-readable summary of the problem
          example: "Bad Request"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "Validation failed for field 'documentId'"
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: "/jobs/abc123"
      required:
        - type
        - title
        - status
  responses:
    BadRequest:
      description: Bad request - validation error or malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "about:blank"
            title: "Bad Request"
            status: 400
            detail: "Validation failed: documentId is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "about:blank"
            title: "Not Found"
            status: 404
            detail: "Job with id 'a1b2c3d4-e5f6-7890-abcd-ef1234567890' not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "about:blank"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing the request"